+++ simulation/gpu/naive_gpu_root_coordinator.cuh
@@ -142,7 +142,7 @@ void gpu_simulation(size_t n_subcomponents, AtomicGPU* subcomponents, size_t* n_
 
 		// Step 2: route messages
 		for(size_t i=0; i<n_subcomponents;i++){
-			for(int j=0; j<n_couplings[i]; j++ ){
+			for(size_t j=0; j<n_couplings[i]; j++ ){
 				subcomponents[i].insert_in_bag(subcomponents[couplings[i][j]].get_out_bag());
 			}
 		}
